перерыв

+++ shop/basketapp/views.py
@@ -1,12 +1,13 @@
 
 from django.shortcuts import render, get_object_or_404, HttpResponseRedirect
 from django.contrib.auth.decorators import login_required
+from django.template.loader import render_to_string
+from django.http import JsonResponse, Http404
 from mainapp.models import Product
 from mainapp.views import get_basket
 from .models import OrderItem
 
-from django.template.loader import render_to_string
-from django.http import JsonResponse, Http404
+
 
 
 # Create your views here.
@@ -19,30 +20,37 @@ def basket_view(request):
 
 @login_required
 def basket_edit(request, pk, value=0):
-    basket = get_basket(user=request.user)
-
-    old_basket_item = get_object_or_404(OrderItem, pk=pk)
-    if old_basket_item:
-        if value == 0:
-            OrderItem.objects.filter(pk=pk).delete()
+    if request.is_ajax():
+        basket = get_basket(user=request.user)
+        print('1')
+        old_basket_item = get_object_or_404(OrderItem, pk=pk)
+        if old_basket_item:
+            print('2')
+            if value == 0:
+                print('3')
+                OrderItem.objects.filter(pk=pk).delete()
+            else:
+                print('4')
+                old_basket_item.quantity = value
+                old_basket_item.save()
         else:
-            old_basket_item.quantity = value
-            old_basket_item.save()
+            print('5')
+            product = get_object_or_404(Product, pk=pk)
+            new_basket_item = OrderItem(basket=basket, product=product)
+            new_basket_item.quantity = 1
+            new_basket_item.save()
+        items = OrderItem.objects.filter(basket=basket)
+        content = {"basket":basket, "items":items,}
+        result = render_to_string('basketapp/basket_inc.html', content)
     else:
-        product = get_object_or_404(Product, pk=pk)
-        new_basket_item = OrderItem(basket=basket, product=product)
-        new_basket_item.quantity = 1
-        new_basket_item.save()
-    items = OrderItem.objects.filter(basket=basket)
-    content = {"basket":basket, "items":items,}
-    result = render_to_string('basketapp/basket_inc.html', content)
+        return Http404
 
 
     return JsonResponse({'result': result})
 
 def basket_remove(request, pk):
     basket = get_basket(user=request.user)
-    print(pk)
+
     product = get_object_or_404(Product, pk=pk)
 
     old_basket_item = OrderItem.objects.filter(basket=basket, product=product)
@@ -54,6 +62,24 @@ def basket_remove(request, pk):
         else:
             old_basket_item[0].quantity -= 1
             old_basket_item[0].save()
+
+    return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
+
+def basket_add(request, pk):
+    basket = get_basket(user=request.user)
+
+    product = get_object_or_404(Product, pk=pk)
+
+    old_basket_item = OrderItem.objects.filter(basket=basket, product=product)
+    if not old_basket_item:
+        product = get_object_or_404(Product, pk=pk)
+        new_basket_item = OrderItem(basket=basket, product=product)
+        new_basket_item.quantity = 1
+        new_basket_item.save()
+    else:
+        print(old_basket_item[0].quantity)
+        old_basket_item[0].quantity += 1
+        print(old_basket_item[0].quantity)
 
     return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
